examples/api/python_client/requirements.txt
requests
python-dotenv
pyyaml 